itests/pyproject.toml
[tool.poetry]
package-mode = false
name = "jupiter-itests"
version = "0.1.0"
description = "Big integration tests"
authors = ["Horia Coman <horia141@gmail.com>"]
license = "MIT"
homepage = "https://get-thriving.com"
readme = "README.md"
repository = "https://github.com/horia141/jupiter"
documentation = "https://docs.get-thriving.com"
# packages = [{ include = "integration" }]

[tool.poetry.dependencies]
python = ">=3.10, <3.11"
jupiter-core = {path = "../src/core", develop = true}
jupiter-webapi-client = {path = "../gen/py/webapi-client", develop = true}
pytest-playwright = "^0.7.0"
playwright = "^1.51.0"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
